database-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cx_Oracle==8.3.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
python-dateutil==2.8.2
openpyxl==3.1.2 